backend/requirements.txt
# Core FastAPI and server
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.5.0
starlette==0.37.0

# Core AI and ML libraries (optional extras)
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data processing and analysis
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
networkx>=3.1.0

# Web framework and API
aiohttp>=3.8.0
requests>=2.31.0

# Database and storage
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# System monitoring
psutil>=5.9.0

# Security and encryption
cryptography>=41.0.0

# Configuration and utilities
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional heavy dependencies (install separately for full features)
# pip install torch transformers qiskit syft